Staging

+++ datafs/core/data_file.py
@@ -44,47 +44,46 @@ def _touch(filesystem, path):
 
 
 @contextmanager
-def _choose_read_fs(authority, cache, service_path, version_check, hasher):
+def _choose_read_fs(authority, cache, read_path, version_check, hasher):
     '''
     Context manager returning the appropriate up-to-date readable filesystem
 
     Use ``cache`` if it is a valid filessystem and has a file at
-    ``service_path``, otherwise use ``authority``. If the file at
-    ``service_path`` is out of date, update the file in ``cache`` before
+    ``read_path``, otherwise use ``authority``. If the file at
+    ``read_path`` is out of date, update the file in ``cache`` before
     returning it.
     '''
 
-    if cache and cache.fs.isfile(service_path):
-        if version_check(hasher(cache.fs.open(service_path, 'rb'))):
+    if cache and cache.fs.isfile(read_path):
+        if version_check(hasher(cache.fs.open(read_path, 'rb'))):
             yield cache.fs
 
-        elif authority.fs.isfile(service_path):
+        elif authority.fs.isfile(read_path):
             fs.utils.copyfile(
                 authority.fs,
-                service_path,
+                read_path,
                 cache.fs,
-                service_path)
+                read_path)
             yield cache.fs
 
         else:
-            _touch(authority.fs, service_path)
-            _touch(cache.fs, service_path)
+            _makedirs(authority.fs, fs.path.dirname(read_path))
+            _makedirs(cache.fs, fs.path.dirname(read_path))
             yield cache.fs
 
     else:
-        if not authority.fs.isfile(service_path):
-            _touch(authority.fs, service_path)
+        if not authority.fs.isfile(read_path):
+            _makedirs(authority.fs, fs.path.dirname(read_path))
 
         yield authority.fs
 
 
 @contextmanager
-def _choose_write_fs(cache, service_path):
+def _get_write_fs():
     '''
     Context manager returning a writable filesystem
 
-    Use the cache if the cache is a valid filessystem and caching is enabled for
-    this archive, otherwise use a temporary directory and clean on exit.
+    Use a temporary directory and clean on exit.
 
     .. todo::
 
@@ -95,40 +94,48 @@ def _choose_write_fs(cache, service_path):
         in-memory filesystem and then writing to both cache and auth.
 
     '''
+    
+    tmp = tempfile.mkdtemp()
 
-    if cache and cache.fs.isfile(service_path):
-
-        yield cache.fs
-
-    else:
-        tmp = tempfile.mkdtemp()
-
+    try:
         # Create a writeFS and path to the directory containing the archive
         write_fs = OSFS(tmp)
 
-        yield write_fs
+        try:
 
-        _close(write_fs)
+            yield write_fs
+
+        finally:
+            _close(write_fs)
+    
+    finally:
         shutil.rmtree(tmp)
 
 
 
 @contextmanager
-def _prepare_write_fs(read_fs, cache, service_path, readwrite_mode=True):
+def _prepare_write_fs(read_fs, cache, read_path, readwrite_mode=True):
+    '''
+    Prepare a temporary filesystem for writing to read_path
 
-    with _choose_write_fs(cache, service_path) as write_fs:
+    The file will be moved to write_path on close if modified.
+    '''
+
+    with _get_write_fs() as write_fs:
 
         # If opening in read/write or append mode, make sure file data is
         # accessible
         if readwrite_mode:
 
-            if not write_fs.isfile(service_path):
-                _touch(write_fs, service_path)
-                fs.utils.copyfile(
-                    read_fs, service_path, write_fs, service_path)
+            if not write_fs.isfile(read_path):
+                _touch(write_fs, read_path)
+
+                if read_fs.isfile(read_path):
+                    fs.utils.copyfile(
+                        read_fs, read_path, write_fs, read_path)
 
         else:
-            _touch(write_fs, service_path)
+            _touch(write_fs, read_path)
 
         yield write_fs
 
@@ -140,9 +147,11 @@ def open_file(
         authority,
         cache,
         update,
-        service_path,
         version_check,
         hasher,
+        read_path,
+        write_path=None,
+        cache_on_write=False,
         mode='r',
         *args,
         **kwargs):
@@ -167,7 +176,10 @@ def open_file(
          update, service_path, version_check, **kwargs
     '''
 
-    with _choose_read_fs(authority, cache, service_path, version_check, hasher) as read_fs:
+    if write_path is None:
+        write_path = read_path
+
+    with _choose_read_fs(authority, cache, read_path, version_check, hasher) as read_fs:
 
         write_mode = ('w' in mode) or ('a' in mode) or ('+' in mode)
 
@@ -178,27 +190,47 @@ def open_file(
                     ('r' in mode) and (
                         '+' in mode)))
 
-            with _prepare_write_fs(read_fs, cache, service_path, readwrite_mode) as write_fs:
+            with _prepare_write_fs(read_fs, cache, read_path, readwrite_mode) as write_fs:
 
                 wrapper = MultiFS()
                 wrapper.addfs('reader', read_fs)
                 wrapper.setwritefs(write_fs)
 
-                with wrapper.open(service_path, mode, *args, **kwargs) as f:
+                with wrapper.open(read_path, mode, *args, **kwargs) as f:
 
                     yield f
 
-                with write_fs.open(service_path, 'rb') as f:
+                with write_fs.open(read_path, 'rb') as f:
                     checksum = hasher(f)
 
                 if not version_check(checksum):
-                    fs.utils.copyfile(
-                        write_fs, service_path, authority.fs, service_path)
-                    update(checksum=checksum)
+                    if (
+                            cache_on_write or 
+                            (
+                                cache 
+                                and (
+                                    fs.path.abspath(read_path) == fs.path.abspath(write_path))
+                                and cache.fs.isfile(read_path)
+                            )
+                        ):
+                        _makedirs(cache.fs, fs.path.dirname(write_path))
+                        fs.utils.copyfile(
+                            write_fs, read_path, cache.fs, write_path)
+
+                        _makedirs(authority.fs, fs.path.dirname(write_path))
+                        fs.utils.copyfile(
+                            cache.fs, write_path, authority.fs, write_path)
+                        
+                    else:
+                        _makedirs(authority.fs, fs.path.dirname(write_path))
+                        fs.utils.copyfile(
+                            write_fs, read_path, authority.fs, write_path)
+
+                    update(**checksum)
 
         else:
 
-            with read_fs.open(service_path, mode, *args, **kwargs) as f:
+            with read_fs.open(read_path, mode, *args, **kwargs) as f:
 
                 yield f
 
@@ -208,9 +240,11 @@ def get_local_path(
         authority,
         cache,
         update,
-        service_path,
         version_check,
-        hasher):
+        hasher,
+        read_path,
+        write_path=None,
+        cache_on_write=False):
     '''
     Context manager for retrieving a system path for I/O and updating on changes
 
@@ -232,23 +266,45 @@ def get_local_path(
          update, service_path, version_check, **kwargs
     '''
 
-    with _choose_read_fs(authority, cache, service_path, version_check, hasher) as read_fs:
+    if write_path is None:
+        write_path = read_path
 
-        with _prepare_write_fs(read_fs, cache, service_path, readwrite_mode=True) as write_fs:
+    with _choose_read_fs(authority, cache, read_path, version_check, hasher) as read_fs:
 
-            yield write_fs.getsyspath(service_path)
+        with _prepare_write_fs(read_fs, cache, read_path, readwrite_mode=True) as write_fs:
 
-            if write_fs.isfile(service_path):
+            yield write_fs.getsyspath(read_path)
 
-                with write_fs.open(service_path, 'rb') as f:
+            if write_fs.isfile(read_path):
+
+                with write_fs.open(read_path, 'rb') as f:
                     checksum = hasher(f)
 
                 if not version_check(checksum):
-                    fs.utils.copyfile(
-                        write_fs, service_path, authority.fs, service_path)
-                    update(checksum=checksum)
+
+                    if (
+                            cache_on_write or 
+                            (
+                                cache 
+                                and (
+                                    fs.path.abspath(read_path) == fs.path.abspath(write_path))
+                                and cache.fs.isfile(read_path)
+                            )
+                        ):
+
+                        _makedirs(cache.fs, fs.path.dirname(write_path))
+                        fs.utils.copyfile(
+                            write_fs, read_path, cache.fs, write_path)
+
+                        _makedirs(authority.fs, fs.path.dirname(write_path))
+                        fs.utils.copyfile(
+                            cache.fs, write_path, authority.fs, write_path)
+                    else:
+                        _makedirs(authority.fs, fs.path.dirname(write_path))
+                        fs.utils.copyfile(
+                            write_fs, read_path, authority.fs, write_path)
+                    update(**checksum)
 
             else:
-                _touch(write_fs, service_path)
                 raise OSError(
                     'Local file removed during execution. Archive not updated.')
